[expo-ui] Add color support for `ContextMenu` components. (#34787)
# Why

The components inside of the `ContextMenu` were using the default
colors. We need them to use the platform theme colors by default and
accept the color props just like when they are used outside of the menu.

# How

All of the changes are for Android, because it is not possible to apply
custom colors inside of the iOS context menu.
Added a `color` prop for the `ContextMenu`, which determines the menu
container color. Components inside of the context menu should now use
the provided color
I reduced the amount of duplicated code between the `Switch` and
`ContextMenu` implementation of the switch.

# Test Plan

Tested in BareExpo on Android

<img width="338" alt="image"
src="https://github.com/user-attachments/assets/53023d79-287b-477b-95c3-974993ea9016"
/>